Move LowLevelLifoSemaphore to System.Threading instead of System.Threading.ClrThreadPool
